+++ inc/template-functions.php
@@ -48,6 +48,31 @@ function secretum_text( $key = '', $echo = false ) {
 
 
 /**
+ * Customizer Refresh Icon
+ *
+ * @since 1.0.0
+ */
+function secretum_customizer_refresh() {
+	echo wp_kses(
+		'<a href="javascript:void(0);" onclick="document.location.reload(true)" title="' . __( 'Refresh Preview', 'secretum' ) . '"><i class="secretum-customizer-icon fi-refresh" aria-hidden="true"></i></a>',
+		[
+			'a' => [
+				'href' 			=> true,
+				'onclick' 		=> true,
+				'title' 		=> true,
+			],
+			'i' => [
+				'class' 		=> true,
+				'aria-hidden' 	=> true,
+			],
+		],
+		'javascript'
+	);
+
+}//end secretum_customizer_refresh()
+
+
+/**
  * Check if WooCommerce is Active
  *
  * @since 1.0.0
@@ -120,9 +145,6 @@ function secretum_breadcrumbs( $taxonomy = '', $top_link = false, $icons = false
 
 	// If Items Set.
 	if ( isset( $terms[0]->name ) && isset( $terms[0]->slug ) ) {
-		// Get Home URL.
-		$home_url = esc_url( home_url() );
-
 		// Get Category Name.
 		$category_name = sanitize_text_field( $terms[0]->name );
 
@@ -130,8 +152,7 @@ function secretum_breadcrumbs( $taxonomy = '', $top_link = false, $icons = false
 		$category_slug = sanitize_html_class( $terms[0]->slug );
 
 		// Build Category URL.
-		// Testing Dont Use $category_url = xget_term_link( $category_slug, sanitize_html_class( $taxonomy ) );.
-		$category_url = '#';
+		$category_url = xget_term_link( $category_slug, sanitize_html_class( $taxonomy ) );
 
 		// Home Icon.
 		$home = ( $icons ) ? '<i class="fa fa-home" aria-hidden="true"></i> ' : '';
@@ -153,10 +174,12 @@ function secretum_breadcrumbs( $taxonomy = '', $top_link = false, $icons = false
 
 		// Build Top Return Link.
 		if ( true === $top_link && true === $icons ) {
-			$top = '<a href="#top" class="ml-2 p-2"><i class="fa fa-caret-up" aria-hidden="true" title="Return to Top"></i></a>';
+			$title = secretum_text( 'return_to_top_title' );
+			$top = '<a href="#top" class="ml-2 p-2"><i class="fa fa-caret-up" aria-hidden="true" title="' . $title . '"></i></a>';
 
 		} elseif ( true === $top_link && false === $icons ) {
-			$top = ' | <a href="#top">top ^</a>';
+			$title = secretum_text( 'return_to_top_default' );
+			$top = ' | <a href="#top">' . $title . '</a>';
 
 		} else {
 			$top = '';
@@ -164,7 +187,7 @@ function secretum_breadcrumbs( $taxonomy = '', $top_link = false, $icons = false
 		}
 
 		// Return HTML.
-		return '<div class="breadcrumbs">' . $home . '<a href="' . $home_url . '">' . $home_text . '</a> ' . $sep . ' <a href="' . $category_url . '">' . $category_name . '</a>' . $top . '</div>';
+		return '<div class="breadcrumbs">' . $home . '<a href="' . SECRETUM_BASE_URL . '">' . $home_text . '</a> ' . $sep . ' <a href="' . $category_url . '">' . $category_name . '</a>' . $top . '</div>';
 	}// End if().
 
 }//end secretum_breadcrumbs()