+++ inc/template-functions/author.php
@@ -21,9 +21,9 @@ namespace Secretum;
 function secretum_author_post_list() {
 	$post_text 	= secretum_text( 'author_posted_text', false );
 	$get_url 	= esc_url( get_permalink() );
-	$get_time 	= esc_attr( get_the_time() );
-	$get_iso 	= esc_attr( get_the_date( 'c' ) );
-	$get_date 	= esc_attr( get_the_date() );
+	$get_time 	= esc_html( get_the_time() );
+	$get_iso 	= esc_html( get_the_date( 'c' ) );
+	$get_date 	= esc_html( get_the_date() );
 	$cat_text 	= secretum_text( 'author_categories_text', false );
 	$cat_list 	= get_the_category_list( ', ' );