+++ woocommerce/order/order-downloads.php
@@ -54,7 +54,7 @@ namespace Secretum;
 							break;
 							case 'download-expires':
 								if ( ! empty( $download['access_expires'] ) ) {
-									echo '<time datetime="' . esc_attr( date( 'Y-m-d', strtotime( $download['access_expires'] ) ) ) . '" title="' . esc_attr( strtotime( $download['access_expires'] ) ) . '">' . esc_html( date_i18n( get_option( 'date_format' ), strtotime( $download['access_expires'] ) ) ) . '</time>';
+									echo '<time datetime="' . esc_html( date( 'Y-m-d', strtotime( $download['access_expires'] ) ) ) . '" title="' . esc_html( strtotime( $download['access_expires'] ) ) . '">' . esc_html( date_i18n( get_option( 'date_format' ), strtotime( $download['access_expires'] ) ) ) . '</time>';
 								} else {
 									esc_html_e( 'Never', 'secretum' );
 								}